refactor: move HID features into plugin projects
