--- DesktopApplicationTemplate.UI/EditHandlers/HidEditServiceHandler.cs
@@ -1,98 +1,0 @@
-using System;
-using DesktopApplicationTemplate.UI.Views;
-using DesktopApplicationTemplate.UI.ViewModels;
-using DesktopApplicationTemplate.UI.ViewModels.Hid.Edit;
-using DesktopApplicationTemplate.UI.ViewModels.Hid.Advanced;
-using DesktopApplicationTemplate.UI.Views.Hid.Edit;
-using DesktopApplicationTemplate.UI.Views.Hid.Advanced;
-using DesktopApplicationTemplate.UI.Services;
-using DesktopApplicationTemplate.Models;
-using DesktopApplicationTemplate.Core.Services;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-
-namespace DesktopApplicationTemplate.UI.EditHandlers;
-
-public class HidEditServiceHandler : IEditServiceHandler
-{
-    private readonly Func<MainView> _getMainView;
-    private readonly Func<MainViewModel> _getMainViewModel;
-    private readonly IServiceProvider _services;
-    private readonly ILogger<HidEditServiceHandler>? _logger;
-
-    public HidEditServiceHandler(Func<MainView> getMainView, Func<MainViewModel> getMainViewModel, IServiceProvider services, ILogger<HidEditServiceHandler>? logger = null)
-    {
-        _getMainView = getMainView;
-        _getMainViewModel = getMainViewModel;
-        _services = services;
-        _logger = logger;
-    }
-
-    public void Edit(ServiceListModel service)
-    {
-        var mainView = _getMainView();
-        var mainViewModel = _getMainViewModel();
-        var hidPage = mainView.GetOrCreateServicePage(service);
-        var payload = service.GetPayload<HidServiceOptions>();
-        if (payload is null)
-        {
-            payload = new HidServiceOptions();
-            service.SetPayload(payload);
-        }
-
-        var vm = ActivatorUtilities.CreateInstance<HidEditServiceViewModel>(_services, service.DisplayName.Split(" - ").Last(), payload);
-        var editView = _services.GetRequiredService<HidEditServiceView>();
-        editView.Initialize(vm);
-        vm.ServiceSaved += (name, opts) =>
-        {
-            var catalog = _services.GetRequiredService<IServiceCatalog>();
-            var descriptor = ResolveDescriptor(catalog, service.DescriptorId, service.Type);
-            service.ApplyDescriptor(descriptor, name);
-            service.SetPayload(opts);
-            if (hidPage != null)
-            {
-                mainView.ShowPage(hidPage);
-            }
-
-            _ = mainViewModel.SaveServicesAsync();
-        };
-        vm.EditCancelled += () =>
-        {
-            if (hidPage != null)
-            {
-                mainView.ShowPage(hidPage);
-            }
-        };
-        vm.AdvancedConfigRequested += opts =>
-        {
-            var advVm = ActivatorUtilities.CreateInstance<HidAdvancedConfigViewModel>(_services, opts);
-            var advView = _services.GetRequiredService<HidAdvancedConfigView>();
-            advView.Initialize(advVm);
-            advVm.Saved += _ => mainView.ShowPage(editView);
-            advVm.BackRequested += () => mainView.ShowPage(editView);
-            mainView.ShowPage(advView);
-        };
-        mainView.ShowPage(editView);
-        _logger?.LogDebug("Edit workflow completed for {Name}", service.DisplayName);
-    }
-
-    private static IServiceDescriptor? ResolveDescriptor(IServiceCatalog catalog, string descriptorId, ServiceType serviceType)
-    {
-        if (!string.IsNullOrWhiteSpace(descriptorId) && catalog.TryGetById(descriptorId, out var byId))
-        {
-            return byId;
-        }
-
-        if (catalog.TryGetByLegacyType(serviceType, out var legacy))
-        {
-            return legacy;
-        }
-
-        if (catalog.LegacyMap.TryGetValue(serviceType, out var fallbackId) && catalog.TryGetById(fallbackId, out var fallback))
-        {
-            return fallback;
-        }
-
-        return null;
-    }
-}
