+++ DesktopApplicationTemplate.UI/Helpers/HookReleaseHelper.cs
@@ -1,5 +1,6 @@
 using System;
-using DesktopApplicationTemplate.UI.Services;
+using DesktopApplicationTemplate.Core.Services;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace DesktopApplicationTemplate.UI.Helpers
 {
@@ -16,11 +17,19 @@ namespace DesktopApplicationTemplate.UI.Helpers
         /// <summary>
         /// Releases any installed HID or keyboard hooks.
         /// </summary>
-        public static void Release()
+        public static void Release(IServiceProvider? serviceProvider = null)
         {
             try
             {
-                KeyboardSimulator.Reset();
+                serviceProvider ??= App.AppHost?.Services;
+                if (serviceProvider is not null)
+                {
+                    var releaseServices = serviceProvider.GetServices<IHookReleaseService>();
+                    foreach (var releaseService in releaseServices)
+                    {
+                        releaseService?.Release();
+                    }
+                }
             }
             finally
             {
