--- DesktopApplicationTemplate.UI/Views/Hid/Create/HidCreateServiceView.xaml.cs
@@ -1,15 +1,0 @@
-using System.Windows.Controls;
-using DesktopApplicationTemplate.Core.Services;
-using DesktopApplicationTemplate.UI.ViewModels.Hid.Create;
-
-namespace DesktopApplicationTemplate.UI.Views.Hid.Create;
-
-public partial class HidCreateServiceView : Page
-{
-    public HidCreateServiceView(HidCreateServiceViewModel vm, ILoggingService logger)
-    {
-        InitializeComponent();
-        DataContext = vm;
-        vm.Logger = logger;
-    }
-}
