--- DesktopApplicationTemplate.UI/Views/Hid/Edit/HidEditServiceView.xaml
@@ -1,43 +1,0 @@
-<Page x:Class="DesktopApplicationTemplate.UI.Views.Hid.Edit.HidEditServiceView"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-      xmlns:views="clr-namespace:DesktopApplicationTemplate.UI.Views"
-      mc:Ignorable="d">
-    <ScrollViewer VerticalScrollBarVisibility="Auto">
-        <Grid Margin="20">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="Auto" />
-                <ColumnDefinition Width="*" />
-            </Grid.ColumnDefinitions>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="Auto" />
-            </Grid.RowDefinitions>
-
-            <TextBlock Grid.Row="0" Grid.Column="0" Text="Service Name" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="0" Grid.Column="1" Text="{Binding ServiceName}" Style="{StaticResource FormField}"/>
-
-            <TextBlock Grid.Row="1" Grid.Column="0" Text="Message Template" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="1" Grid.Column="1" Text="{Binding MessageTemplate}" Style="{StaticResource FormField}" Height="80" AcceptsReturn="True"/>
-
-            <TextBlock Grid.Row="2" Grid.Column="0" Text="USB Protocol" Style="{StaticResource FormLabel}"/>
-            <ComboBox Grid.Row="2" Grid.Column="1" ItemsSource="{Binding UsbProtocols}" SelectedItem="{Binding SelectedUsbProtocol}" Style="{StaticResource FormField}"/>
-
-            <TextBlock Grid.Row="3" Grid.Column="0" Text="Attached Service" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="3" Grid.Column="1" Text="{Binding AttachedService}" Style="{StaticResource FormField}"/>
-
-            <views:EditorButtonBar Grid.Row="4" Grid.ColumnSpan="2"
-                                   AdvancedConfigCommand="{Binding AdvancedConfigCommand}"
-                                   SaveCommand="{Binding SaveCommand}"
-                                   CancelCommand="{Binding CancelCommand}"
-                                   AdvancedAutomationName="Open HID Advanced Configuration"
-                                   SaveAutomationName="Save HID Service"
-                                   CancelAutomationName="Cancel HID Edit" />
-        </Grid>
-    </ScrollViewer>
-</Page>
