--- DesktopApplicationTemplate.UI/Views/Hid/Edit/HidEditServiceView.xaml.cs
@@ -1,22 +1,0 @@
-using System.Windows.Controls;
-using DesktopApplicationTemplate.Core.Services;
-using DesktopApplicationTemplate.UI.ViewModels.Hid.Edit;
-
-namespace DesktopApplicationTemplate.UI.Views.Hid.Edit;
-
-public partial class HidEditServiceView : Page
-{
-    private readonly ILoggingService _logger;
-
-    public HidEditServiceView(ILoggingService logger)
-    {
-        InitializeComponent();
-        _logger = logger;
-    }
-
-    public void Initialize(HidEditServiceViewModel vm)
-    {
-        DataContext = vm;
-        vm.Logger = _logger;
-    }
-}
