+++ DesktopApplicationTemplate.sln
@@ -37,6 +37,10 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FtpUI", "Services\Ftp\FtpUI\FtpUI.csproj", "{389FAA88-2E7D-4093-ABB1-F6AAB6D15E38}"
 EndProject
 
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HidPackage", "Services\Hid\HidPackage\HidPackage.csproj", "{8D51DB36-0D19-4B75-8566-C33E410694F8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HidUI", "Services\Hid\HidUI\HidUI.csproj", "{4A30CC96-8FD1-4298-9888-D9BCDF793A3F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -107,6 +111,14 @@ Global
                 {389FAA88-2E7D-4093-ABB1-F6AAB6D15E38}.Debug|Any CPU.Build.0 = Debug|Any CPU
                 {389FAA88-2E7D-4093-ABB1-F6AAB6D15E38}.Release|Any CPU.ActiveCfg = Release|Any CPU
                 {389FAA88-2E7D-4093-ABB1-F6AAB6D15E38}.Release|Any CPU.Build.0 = Release|Any CPU
+                {8D51DB36-0D19-4B75-8566-C33E410694F8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {8D51DB36-0D19-4B75-8566-C33E410694F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {8D51DB36-0D19-4B75-8566-C33E410694F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {8D51DB36-0D19-4B75-8566-C33E410694F8}.Release|Any CPU.Build.0 = Release|Any CPU
+                {4A30CC96-8FD1-4298-9888-D9BCDF793A3F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {4A30CC96-8FD1-4298-9888-D9BCDF793A3F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {4A30CC96-8FD1-4298-9888-D9BCDF793A3F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {4A30CC96-8FD1-4298-9888-D9BCDF793A3F}.Release|Any CPU.Build.0 = Release|Any CPU
         EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
