+++ Services/Hid/HidPackage/HidPackage.csproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <Import Project="..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.props" Condition="Exists('..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.props')" />
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <AssemblyName>DesktopApplicationTemplate.Hid</AssemblyName>
+    <RootNamespace>DesktopApplicationTemplate.Services.Hid</RootNamespace>
+    <ServicePluginPackageId>DesktopApplicationTemplate.Hid</ServicePluginPackageId>
+    <ServicePluginPackageVersion>1.0.0</ServicePluginPackageVersion>
+  </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\DesktopApplicationTemplate.Core\DesktopApplicationTemplate.Core.csproj" />
+    <ProjectReference Include="..\..\..\DesktopApplicationTemplate.Services.Common\DesktopApplicationTemplate.Services.Common.csproj" />
+    <ProjectReference Include="..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.csproj" ReferenceOutputAssembly="false" />
+  </ItemGroup>
+  <Import Project="..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.targets" Condition="Exists('..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.targets')" />
+</Project>
