+++ Services/Hid/HidUI/Navigation/HidNavigationHandler.cs
@@ -1,0 +1,60 @@
+using System;
+using System.Threading.Tasks;
+using System.Windows.Controls;
+using DesktopApplicationTemplate.Core.Services;
+using DesktopApplicationTemplate.Services.Hid.Descriptors;
+using DesktopApplicationTemplate.Services.Hid.UI.ViewModels.Hid.Advanced;
+using DesktopApplicationTemplate.Services.Hid.UI.ViewModels.Hid.Create;
+using DesktopApplicationTemplate.Services.Hid.UI.Views.Hid.Advanced;
+using DesktopApplicationTemplate.Services.Hid.UI.Views.Hid.Create;
+using DesktopApplicationTemplate.UI.Factories;
+using DesktopApplicationTemplate.UI.Views;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace DesktopApplicationTemplate.Services.Hid.UI.Navigation
+{
+    [ServiceDescriptorRegistration(HidServiceDescriptor.DescriptorId, ServiceRegistrationKind.NavigationHandler)]
+    public class HidNavigationHandler : INavigationHandler
+    {
+        private readonly IServiceProvider _services;
+        private readonly Func<MainView> _getMainView;
+        private readonly IServiceCatalog _catalog;
+
+        public string DescriptorId => HidServiceDescriptor.DescriptorId;
+
+        public HidNavigationHandler(IServiceProvider services, Func<MainView> getMainView, IServiceCatalog catalog)
+        {
+            _services = services;
+            _getMainView = getMainView;
+            _catalog = catalog;
+        }
+
+        public Page CreateView(string defaultName)
+        {
+            var vm = _services.GetRequiredService<HidCreateServiceViewModel>();
+            vm.ServiceName = defaultName;
+            var mainView = _getMainView();
+            vm.ServiceSaved += (name, options) => _ = AddServiceAsync(name, options);
+            vm.EditCancelled += mainView.ShowCreateServiceSelectionPage;
+            var view = ActivatorUtilities.CreateInstance<HidCreateServiceView>(_services, vm);
+            vm.AdvancedConfigRequested += opts =>
+            {
+                var advVm = ActivatorUtilities.CreateInstance<HidAdvancedConfigViewModel>(_services, opts);
+                var advView = _services.GetRequiredService<HidAdvancedConfigView>();
+                advView.Initialize(advVm);
+                advVm.Saved += _ => mainView.ShowPage(view);
+                advVm.BackRequested += () => mainView.ShowPage(view);
+                mainView.ShowPage(advView);
+            };
+            return view;
+        }
+
+        public Task AddServiceAsync(string name, object options)
+        {
+            var mainView = _getMainView();
+            _catalog.TryGetById(DescriptorId, out var descriptor);
+            var context = new ServiceFactoryContext(DescriptorId, name, options, descriptor);
+            return mainView.AddServiceAsync(context);
+        }
+    }
+}
