+++ Services/Hid/HidUI/ViewModels/Hid/HidViewModel.cs
@@ -1,0 +1,145 @@
+using System;
+using System.Collections.ObjectModel;
+using System.Windows.Input;
+using DesktopApplicationTemplate.Services.Hid.UI.Services;
+using DesktopApplicationTemplate.UI.Helpers;
+using DesktopApplicationTemplate.UI.Services;
+using DesktopApplicationTemplate.Core.Services;
+
+namespace DesktopApplicationTemplate.Services.Hid.UI.ViewModels.Hid
+{
+    public class HidViewModel : ViewModelBase, IDisposable
+    {
+        private string _messageTemplate = string.Empty;
+        public string MessageTemplate
+        {
+            get => _messageTemplate;
+            set { _messageTemplate = value; OnPropertyChanged(); }
+        }
+
+        public ObservableCollection<string> AvailableServices { get; } = new();
+
+        private string _attachedService = string.Empty;
+        public string AttachedService
+        {
+            get => _attachedService;
+            set { _attachedService = value; OnPropertyChanged(); Logger?.Log($"Attached service set to {value}", LogLevel.Debug); }
+        }
+
+        public ObservableCollection<string> UsbProtocols { get; } = new() { "2.0", "3.0" };
+
+        private string _selectedUsbProtocol = "2.0";
+        public string SelectedUsbProtocol
+        {
+            get => _selectedUsbProtocol;
+            set { _selectedUsbProtocol = value; OnPropertyChanged(); Logger?.Log($"USB protocol set to {value}", LogLevel.Debug); }
+        }
+
+        private string _debounceTimeMs = "0";
+        public string DebounceTimeMs
+        {
+            get => _debounceTimeMs;
+            set { _debounceTimeMs = value; OnPropertyChanged(); Logger?.Log($"Debounce time set to {value}ms", LogLevel.Debug); }
+        }
+
+        private string _keyDownTimeMs = "0";
+        public string KeyDownTimeMs
+        {
+            get => _keyDownTimeMs;
+            set { _keyDownTimeMs = value; OnPropertyChanged(); Logger?.Log($"Key down time set to {value}ms", LogLevel.Debug); }
+        }
+
+        private string _formatTemplate = "{0}";
+        public string FormatTemplate
+        {
+            get => _formatTemplate;
+            set { _formatTemplate = value; OnPropertyChanged(); }
+        }
+
+        public ILoggingService? Logger { get; set; }
+
+        private string _finalMessage = string.Empty;
+        public string FinalMessage
+        {
+            get => _finalMessage;
+            set { _finalMessage = value; OnPropertyChanged(); }
+        }
+
+        private string _incomingData = string.Empty;
+        /// <summary>Raw data received before processing.</summary>
+        public string IncomingData
+        {
+            get => _incomingData;
+            set { _incomingData = value; OnPropertyChanged(); }
+        }
+
+        private string _processingData = string.Empty;
+        /// <summary>Intermediate representation used during transformations.</summary>
+        public string ProcessingData
+        {
+            get => _processingData;
+            set { _processingData = value; OnPropertyChanged(); }
+        }
+
+        private string _outgoingData = string.Empty;
+        /// <summary>Resulting data after processing.</summary>
+        public string OutgoingData
+        {
+            get => _outgoingData;
+            set { _outgoingData = value; OnPropertyChanged(); }
+        }
+
+        public ICommand BuildCommand { get; }
+        public ICommand SaveCommand { get; }
+
+        private readonly SaveConfirmationHelper _saveHelper;
+
+        public HidViewModel(SaveConfirmationHelper saveHelper)
+        {
+            _saveHelper = saveHelper;
+            BuildCommand = new RelayCommand(BuildMessage);
+            SaveCommand = new RelayCommand(Save);
+        }
+
+        private void BuildMessage()
+        {
+            Logger?.Log("Building HID message", LogLevel.Debug);
+            IncomingData = MessageTemplate;
+            ProcessingData = FormatTemplate ?? "{0}";
+            try
+            {
+                FinalMessage = string.Format(ProcessingData, IncomingData);
+                OutgoingData = FinalMessage;
+                Logger?.Log($"Final HID message: {FinalMessage}", LogLevel.Debug);
+                if (!string.IsNullOrWhiteSpace(AttachedService))
+                {
+                    Logger?.Log($"Forwarding message to {AttachedService}", LogLevel.Debug);
+                    MessageForwarder.Forward(AttachedService, FinalMessage);
+                }
+            }
+            catch (FormatException ex)
+            {
+                Logger?.Log($"HID message formatting failed: {ex.Message}", LogLevel.Error);
+                return;
+            }
+            finally
+            {
+                KeyboardSimulator.Reset();
+            }
+        }
+
+        private void Save()
+        {
+            Logger?.Log("Saving HID configuration", LogLevel.Debug);
+            _saveHelper.Show();
+        }
+
+        /// <summary>
+        /// Releases any simulated key presses.
+        /// </summary>
+        public void Dispose()
+        {
+            KeyboardSimulator.Reset();
+        }
+    }
+}
