+++ Services/Hid/HidUI/Views/Hid/Advanced/HidAdvancedConfigView.xaml
@@ -1,0 +1,31 @@
+<Page x:Class="DesktopApplicationTemplate.Services.Hid.UI.Views.Hid.Advanced.HidAdvancedConfigView"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:views="clr-namespace:DesktopApplicationTemplate.UI.Views"
+      mc:Ignorable="d">
+    <Grid Margin="20">
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="Auto" />
+            <ColumnDefinition Width="*" />
+        </Grid.ColumnDefinitions>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+        </Grid.RowDefinitions>
+
+        <TextBlock Grid.Row="0" Grid.Column="0" Text="Debounce (ms)" Style="{StaticResource FormLabel}"/>
+        <TextBox Grid.Row="0" Grid.Column="1" Text="{Binding DebounceTimeMs}" Style="{StaticResource FormField}"/>
+
+        <TextBlock Grid.Row="1" Grid.Column="0" Text="Key Down (ms)" Style="{StaticResource FormLabel}"/>
+        <TextBox Grid.Row="1" Grid.Column="1" Text="{Binding KeyDownTimeMs}" Style="{StaticResource FormField}"/>
+
+        <views:AdvancedConfigButtonBar Grid.Row="2" Grid.ColumnSpan="2"
+                                       SaveCommand="{Binding SaveCommand}"
+                                       BackCommand="{Binding BackCommand}"
+                                       SaveAutomationName="Save HID Advanced Configuration"
+                                       BackAutomationName="Back to HID Configuration"/>
+    </Grid>
+</Page>
