+++ Services/Hid/HidUI/Views/Hid/HidViews.xaml
@@ -1,0 +1,55 @@
+<Page x:Class="DesktopApplicationTemplate.Services.Hid.UI.Views.Hid.HidViews"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:DesktopApplicationTemplate.UI.Views"
+      xmlns:behaviors="clr-namespace:DesktopApplicationTemplate.UI.Behaviors"
+      mc:Ignorable="d"
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="HidViews">
+
+    <Grid Margin="20">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="*"/>
+            <RowDefinition Height="Auto"/>
+        </Grid.RowDefinitions>
+
+        <TextBlock Text="HID Views" FontSize="18" FontWeight="Bold" Margin="0,0,0,10"/>
+
+        <local:DataFlowDiagram Grid.Row="1" Margin="0,0,0,10" />
+
+        <StackPanel Grid.Row="2" Margin="0,20,0,0" Width="400">
+            <TextBlock Text="HID Message" FontWeight="Bold" Margin="0,0,0,5"/>
+            <TextBox Text="{Binding MessageTemplate}" Height="80" AcceptsReturn="True" x:Name="TemplateBox"
+                     behaviors:TextBoxHintBehavior.AutoToolTip="True"/>
+            <TextBlock Text="Format Template" FontWeight="Bold" Margin="0,10,0,5"/>
+            <TextBox Text="{Binding FormatTemplate}" Height="25"
+                     behaviors:TextBoxHintBehavior.AutoToolTip="True"/>
+            <TextBlock Text="Final Message" FontWeight="Bold" Margin="0,10,0,5"/>
+            <TextBox Text="{Binding FinalMessage}" IsReadOnly="True" Height="80"
+                     behaviors:TextBoxHintBehavior.AutoToolTip="True"/>
+            <StackPanel Orientation="Horizontal" Margin="0,10,0,0">
+                <TextBlock Text="USB:" VerticalAlignment="Center" Margin="0,0,5,0"/>
+                <ComboBox Width="80" ItemsSource="{Binding UsbProtocols}" SelectedItem="{Binding SelectedUsbProtocol}"/>
+                <TextBlock Text="Debounce (ms):" VerticalAlignment="Center" Margin="10,0,5,0"/>
+                <TextBox Width="60" Text="{Binding DebounceTimeMs}"
+                         behaviors:TextBoxHintBehavior.AutoToolTip="True"/>
+                <TextBlock Text="Key Down (ms):" VerticalAlignment="Center" Margin="10,0,5,0"/>
+                <TextBox Width="60" Text="{Binding KeyDownTimeMs}"
+                         behaviors:TextBoxHintBehavior.AutoToolTip="True"/>
+            </StackPanel>
+            <StackPanel Orientation="Horizontal" Margin="0,10,0,0">
+                <TextBlock Text="Forward To:" VerticalAlignment="Center" Margin="0,0,5,0"/>
+                <ComboBox Width="200" ItemsSource="{Binding AvailableServices}" SelectedItem="{Binding AttachedService}"/>
+            </StackPanel>
+            <Button Content="Build" Command="{Binding BuildCommand}" Width="100" Margin="0,10,0,0" HorizontalAlignment="Left"/>
+        </StackPanel>
+
+        <StackPanel Grid.Row="3" HorizontalAlignment="Right" Margin="0,10,0,0">
+            <Button Content="Save Configuration" Command="{Binding SaveCommand}" Width="150"/>
+        </StackPanel>
+    </Grid>
+</Page>
