+++ Services/Hid/HidUI/plugin.manifest.json
@@ -1,0 +1,12 @@
+{
+  "id": "DesktopApplicationTemplate.Hid.UI",
+  "name": "HID Service UI",
+  "version": "1.0.0",
+  "entryAssembly": "DesktopApplicationTemplate.Hid.UI.dll",
+  "serviceAssemblies": [
+    "DesktopApplicationTemplate.Hid.UI.dll"
+  ],
+  "probingPaths": [
+    "libs"
+  ]
+}
